FIX caliper-php lib can't be installed
The caliper-php repo went private. So I've switched the source to an
archive.

Found out that you can merge in extension composer.json instead of
duplicating it in composer.local.json, so I changed it to pull in the
caliper-php lib via the caliper extension's composer.json.

I tried merging in all extension composer.json, but it didn't work,
there was a conflict with the mediawiki/maps package, open issue here:
https://github.com/ProfessionalWiki/Maps/issues/745

If it weren't for that, we could use the mediawiki suggested:

{
    "extra": {
        "merge-plugin": {
            "include": [
                "extensions/*/composer.json",
                "skins/*/composer.json"
            ]
        }
    }
}

Which pulls in all extension/skin dependencies automatically.