FIX PluggableAuth6 requires new config to work
PluggableAuth 6+ requires the new var $wgPluggableAuth_Config to be
properly set.

Due to associated changes in LDAPAuthentication2, there doesn't seem to
be a way for the UBCAuth extension to determine the currently active
LDAP domain. So I'm setting $ubcLDAPDomain to the LDAP_DOMAIN env var
and assuming that's the only active LDAP domain we'll need to worry
about.