FIX Science:Math_Exam_Resources DPL error
Science:Math_Exam_Resources DPL errors out with "Column 'rev_actor' in
field list is ambiguous". This appears to be due to the revision table
is getting two aliases: 'rev' and 'revision', resulting in this
ambiguity.  The revision alias seems to be used only in
Query::_minoredits().  Changing Query::_minoredits() to also use the rev
alias seems to fix it for our use case.

It looks like 'rev_actor' was added without a table alias because
Query::buildAndSelect() calls Query::_adduser() via a variable function
on line 184. The other potential fix was to default _adduser() to the
'rev' table alias, but that seems to have the potential for wider
ranging consequences than the change to _minoredits().

Hopefully upstream can provide a better fix, then we can remove our
hack. Upstream issue:
https://github.com/Universal-Omega/DynamicPageList3/issues/238