FIX caliper error when wiki in read-only mode
Update caliper extension to v2.0.5 for the fix.

Added CALIPER_USE_JOB_QUEUE env var to control whether to use the
mediawiki job queue or not. If we want to have caliper enabled while the
wiki is in an extended read-only mode, we can disable the job queue,
this'll force caliper to send events immediately on every page load.
There'll be no retries in this mode, so any failed sends will just be
lost.