EDIT $wgLogo deprecated, now need to use $wgLogos
The old $wgLogo config is deprecated. The new $wgLogos config allows you
define many logo variations. Most of it is for backward compatibility.
If you want to add a logo, the new MEDIAWIKI_LOGO_ICON env var is
required as it is the only one used by new themes, all the other values
will be deprecated in the future and so are optional.

Since we still have the old theme available as an option for users
to pick, we'll still use some of the backward compat logo values:

    MEDIAWIKI_LOGO_LEGACY1X: 135px wide, up to 155px tall
    MEDIAWIKI_LOGO_LEGACY2X: 270px wide, up to 310px tall

I'm not sure if both is required, but I got odd results when only one or
the other is present, so putting both in to be safe. We only have the 2x
version of the logo previously, so will have to upload a 1x version.