HACK more visible login button upper-right corner
Partial backport of Mediawiki fix for bug T289212 in the Vector skin.

The full bugfix moves the login button out of the upper right '...' menu
to make it more visible. This fix seems to be present in REL1_40.
However, the full commit didn't merge in cleanly with our REL1_39 Vector
skin. It looks like it requires a bunch of other changes that I would
rather not bring in, as it probably has some dependency on Mediawiki
REL1_40.

So I've only done a partial backport, which results in two login links.
One in the upper right corner and one still inside the '...' menu.

This hack can safely removed when we upgrade to REL1_40+