HACK Vector skin hack to hide redundant login btn
The user links menu ("..." on the upper right) has a redundant login
button inside it. There's a login button in the overflow menu next to it
that's more visible. I've set this login button to be uncollapsible, so
it should always be visible even on small screens. So we just need to
hide the user links menu for unsigned in users.