EDIT vector skin dropdown menu style main menu
Current Wikipedia's hamburger menu (the main menu) is a dropdown style
that pops over the page content. This takes up less space on smaller
screens than our 1.39 version's main menu which expands to 100% width
and displace the page content when it appears. Wiki team requested that
we try to have something more like current Wikipedia's.

I was unable to backport this feature as there was A LOT of changes to
the Vector skin. So I tried looking for minimal changes to get something
close to the desired behaviour. After quite a bit experimentation, I
found that some CSS mods can at least convert the main menu to a
dropdown. Though it's not as nice as the official upstream's. Of
particular note is that you can dismiss the upstream's dropdown by
clicking elsewhere in the page.

The CSS changes are standard dropdown menu stuff. The aesthetic
borders/transition/width, I copied from upstream.

Turns out the main menu is somewhat coupled with the table of content.
So I had to make some changes to decouple them. In TableOfContents.less,
the changes stop the TOC from jumping when you open the main menu. in
screen.less, the change stops a blank sidebar from being created when
you open the main menu on TOC-less pages.