FIX VisualEditor error when embed certain video
Embedding the Hunga Tonga-Hunga Ha'apai volcano video seems to trigger
this for some reason, source below:

    [[File:Hunga Tonga–Hunga Ha'apai volcanic eruption captured at December 30, 2021.webm|thumb|test]]

When debug is on, the error thrown is "TypeError:
DOMElement::setAttribute() expects parameter 2 to be string, null
given". Found that this was patched in 1.40 on the Wikimedia bug
tracker: https://phabricator.wikimedia.org/T350594

The same fix works cleanly on our 1.39, so I've just added the patch to
the repo and made the Dockerfile apply it.