ADD main menu hide btn, hidden on page navigate
Added a 'hide' button on the top right corner of the main menu that can
be clicked to hide the main menu. People were confused about how to hide
the main menu so this provides a more obvious alternative than clicking
the hamburger button again.

Also made it so that the main menu is hidden everytime someone navigates
to a new page. This is done by removing the persistent state that stores
whether the menu is opened from the checkbox that controlled the main
menu's open/close. This defaults the menu to closed everytime a page is
loaded.