ADD simplesamlphp extension support
SimpleSAMLphp extension requires a SimpleSAMLphp SP application to be
run separately. It also expects the SimpleSAMLphp SP to be on the same
server and will import from the SP. To duplicate this, I'm running the
SP as a separate container but the codebase is in a shared volume that
is mounted in both the wiki and the SP container.

To make the shared volume work, the SP container initially has its
codebase in /var/www/simplesamlphp-base which will be copied into the
shared volume /var/www/simplesamlphp on startup.

There is a local IDP counterpart that is only used for dev testing. In
order to make setup easy, I've committed a set of certs of the SP in
docker/simplesamlphp/sp/cert/ which is mounted into the SP. Warning:
These certs are meant only for local dev and must not be used in any
kind of prod environment.

Check README for details on the SP's env vars and how to generate your
own cert/key.

Note that the SP is using an older version of SimpleSAMLphp that still
supports php7.4. Newer versions require php8 or above and cannot be used
with our current wiki build as its using php7.4.

The SimpleSAMLphp extension adds its own logout button. The default
logout button ends up being a non-functional duplicate. I think the
extension is supposed to remove this duplicate logout button but I get
the feeling the method they use only works for newer vector skin. I've
manually deleted the duplicate logout button.

Configuration of the SimpleSAMLphp extension is largely the same as it's
also a PluggableAuth module. The main difference is that we configure
the SAML attribute mapping there. Some of the attributes are used only
by UBCAuth.

Minor Mediawiki version upgrade 1.39.7 to 1.39.8.

Remove deprecated LDAP hooks in CustomHooks.php. I think the remaining
Caliper hook should only activate if the UBCAuth extension is enabled.