FIX simplesamlphp sp can't bind to port 80 in kube
Only happened when I tried to deploy on kubernetes. Might be because
we're running as www-data and not root, so it doesn't have permission.
This wasn't an issue with docker compose because local docker apparently
treats all ports as unprivileged.

Changed the sp port to 8080, this is mainly apache configuration and
adjusting traefik.