FIX login loop due to unsecured cookie
I think the login loop was due to the browser not being happy with the
cookie. Toggling the secure cookie flag seems to fix this. I also forced
samesite to none at the time, but I'm not sure if that's necessary. This
patch doesn't force samesite to none, cause I'm hoping the automatic
samesite none code works fine without us having to do anything.

So I added code that'll set the secure cookie flag to true if we're
using a SIMPLESAMLPHP_BASEURL that starts with https.

Note that SIMPLESAMLPHP_BASEURL config is needed because SimpleSAMLphp
errors out if the application baseURL isn't https. Since we have an
ingress load balancer in front of the pods handling https, SimpleSAMLphp
itself doesn't know we're actually using https without this setting.

Sent SimpleSAMLphp logs to stderr so they show up in the kubectl logs.
Easiest way for us to see those logs.