FIX logout message security error from idp
" The request cannot be fulfilled because the message received does not
meet the security requirements of the login service. " I think this
means we need to sign our logout messages.

I think this only happens because its trying to do a singlelogout.
While I'd rather just disable singlelogout, there doesn't seem to be a
simple toggle to do this. So I'm trying to get slo working instead.