FIX puid not being stored in cwl data
The OID used for the puid was wrong. So I've set it to the right one.
I've also updated the dev IDP so that it'll send out the puid attribute
on that OID.

I've also changed the eduPersonAffiliation to values that I see in
staging.

I noticed that the dev IDP doesn't have the SP metadata checked in, so
I've added it in.

Also noticed that nodeservices is probably very outdated, although it
still seems to work. Added comments that it might not be necessary in
the next LTS due to deprecations. Couldn't get it to work in dev, but
seems to be working in prod, so not touching it right now.