EDIT LocalSetting enable pluggableauth fast logout
Without fast logout, the user clicks logout and then has to click an
additional form submit button to send the actual logout request to the
IDP. With fast logout, this form submit is sent with a background post
request. The docs note that fast logout might not with some SLO
implementations, but it doesn't seem to have an issue with our staging
CWL, so I'm enabling it.

Increased the key size for the openssl generation command in README.md
to 4096 for a bit more future proofing.