FIX saml login error if existing session elsewhere
If the IDP knows you're already logged in on another site, it'll
authorize you for access without prompting you for the username and
password again. Users in this situation gets properly logged in but will
end up seeing an error message in the wiki about a fatal error. Note
that the login didn't actually fail, if you refresh, you'll see that
you're properly logged in.

Turns out this error was due to the UBCAuth extension not pulling the
REL1_43 version, so it was still getting the database object in a hard
deprecated way. Previously, it was just pulling master and that works as
long as master is the latest stable. I've switched this to explicitly
pull the tagged REL1_43 version, pulling a pinned version is safer for
prod builds anyway.