UPGRADE simplesamlphp to current stable 2.4.2
The SP was using an old version of php that was still compatible with
php7 due to mediawiki 1.39 being still on php7 (simplesamlphp gets
imported by the pluggableauth simplesamlphp extension, so they needed to
be compatible). With the mediawiki upgrade, we can now ditch php7 and go
to php8.3 like the wiki itself.

Perhaps due to some directory/file changes, there was an issue copying
the new simplesamlphp files to the shared volume, so I've added a
command to delete previously copied simplesamlphp files in the shared
volume.

This file deletion ran into an issue where the container user did not
have permission to delete the certs/ directory mounted by docker
compose, so I just changed the mountpoint to simplesamlephp-base/, where
it'll get copied to the shared volume just like everything else.